♻️ change mod structure
